add validation

+++ ex01/RPN.hpp
@@ -9,13 +9,13 @@
 class RPN
 {
 private:
-    std::stack<int> _num_stack;
+    std::stack<long> _num_stack;
 public:
     RPN();
     ~RPN();
     RPN(const RPN &other);
     RPN &operator=(const RPN &other);
-    void push(int num);
+    void push(long num);
     void calc(char ope);
     bool isEnoughSize();
     void displayResult();
